KickoutMemberTest
